Consolidate the nine boolean2 engine headers into src/boolean2.h
