+++ src/boolean2.cpp
@@ -21,9 +21,6 @@
 #include <utility>
 #include <vector>
 
-#include "cross_section/boolean2/driver.h"
-#include "cross_section/boolean2/predicates.h"
-#include "cross_section/boolean2/winding_filter.h"
 #include "manifold/optional_assert.h"
 
 namespace manifold {
