+++ src/boolean2_offset.cpp
@@ -17,7 +17,6 @@
 #include <vector>
 
 #include "boolean2.h"
-#include "cross_section/boolean2/predicates.h"
 #include "utils.h"
 
 // Polygon offset backing `CrossSection::Offset`.
