+++ src/boolean2_winding.cpp
@@ -33,10 +33,8 @@
 #include <utility>
 #include <vector>
 
+#include "boolean2.h"
 #include "boolean2_diagnostics.h"
-#include "cross_section/boolean2/canonicalize.h"
-#include "cross_section/boolean2/predicates.h"
-#include "cross_section/boolean2/winding_filter.h"
 #include "manifold/optional_assert.h"
 #include "parallel.h"
