+++ src/cross_section/boolean2/canonicalize.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "canonicalize.h"
+#include "../../boolean2.h"
 
 namespace manifold {
