--- src/cross_section/boolean2/parallel_policy.h
@@ -1,25 +1,0 @@
-// Copyright 2026 The Manifold Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-namespace manifold {
-
-// Shared grain size for boolean2 parallel loops whose per-item work is small
-// but non-trivial: BVH traversal, pair narrow tests, and merge broad-phase
-// sweeps. Radix-tree construction uses a coarser build-specific grain because
-// its per-index work is much smaller.
-inline constexpr int kFineParallelGrainSize = 512;
-
-}  // namespace manifold
