--- src/cross_section/boolean2/winding_filter.h
@@ -1,36 +1,0 @@
-// Copyright 2026 The Manifold Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <vector>
-
-#include "canonicalize.h"
-
-namespace manifold {
-
-struct Trace;
-
-enum class WindRule {
-  Add,
-  Intersect,
-};
-
-std::vector<OutEdge> FilterByWindingHalfedges(const CanonicalSubEdges& canon,
-                                              const std::vector<vec2>& verts,
-                                              bool debug = false,
-                                              WindRule rule = WindRule::Add,
-                                              Trace* trace = nullptr);
-
-}  // namespace manifold
